Week7-RAG/requirements.txt
faiss-cpu 
sentence-transformers 
transformers 
torch 
pypdf 
python-docx 
pandas 
pyyaml 
tqdm